+++ game_patch/object/object.h
@@ -41,6 +41,8 @@ void riot_shield_do_frame();
 void riot_shield_on_player_spawn(rf::Player* player);
 void riot_shield_reset_fp_decals(rf::Player* player);
 void riot_shield_on_multi_level_init();
+void entity_rate_limit_on_entity_delete(int handle);
+void entity_rate_limit_clear();
 
 constexpr size_t old_obj_limit = 1024;
 constexpr size_t obj_limit = 65536;